Make background tooltip as individual component.
