+++ osu.Game.Rulesets.Karaoke/Graphics/Cursor/SingerToolTip.cs
@@ -1,62 +1,30 @@
 ﻿// Copyright (c) andy840119 <andy840119@gmail.com>. Licensed under the GPL Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-using osu.Framework.Allocation;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
-using osu.Framework.Graphics.Cursor;
 using osu.Framework.Graphics.Shapes;
-using osu.Game.Graphics;
 using osu.Game.Rulesets.Karaoke.Beatmaps.Metadatas;
 using osuTK;
 using osuTK.Graphics;
 
 namespace osu.Game.Rulesets.Karaoke.Graphics.Cursor
 {
-    public class SingerToolTip : VisibilityContainer, ITooltip
+    public class SingerToolTip : BackgroundToolTip
     {
-        private readonly Box background;
-
-        public void Move(Vector2 pos) => Position = pos;
-
         public SingerToolTip()
         {
-            AutoSizeAxes = Axes.Both;
-            Masking = true;
-            CornerRadius = 5;
-
             Children = new Drawable[]
             {
-                background = new Box
+                new Box
                 {
-                    RelativeSizeAxes = Axes.Both
-                },
-                new Container
-                {
-                    AutoSizeAxes = Axes.Both,
-                    AutoSizeDuration = 200,
-                    AutoSizeEasing = Easing.OutQuint,
-
-                    Padding = new MarginPadding(10),
-                    Children = new Drawable[]
-                    {
-                        new Box
-                        {
-                            Size = new Vector2(100),
-                            Colour = Color4.Red,
-                        }
-                    }
+                    Size = new Vector2(100),
+                    Colour = Color4.Red,
                 }
             };
         }
 
-        [BackgroundDependencyLoader]
-        private void load(OsuColour colours)
-        {
-            background.Colour = colours.Gray3;
-        }
-
-        public bool SetContent(object content)
+        public override bool SetContent(object content)
         {
             if(!(content is Singer singer))
                 return false;
@@ -64,9 +32,5 @@ namespace osu.Game.Rulesets.Karaoke.Graphics.Cursor
             // todo : implement
             return true;
         }
-
-        protected override void PopIn() => this.FadeIn(200, Easing.OutQuint);
-
-        protected override void PopOut() => this.FadeOut(200, Easing.OutQuint);
     }
 }
